Factor profile builders into separate modules
