+++ config/profiles/__init__.py
@@ -1,0 +1,1 @@
+"""Profile configuration builders."""
